+++ extensions/quick-access/src/components/quick-access-empty-view.tsx
@@ -1,6 +1,6 @@
 import { Action, ActionPanel, Icon, List } from "@raycast/api";
 import React, { Dispatch, SetStateAction } from "react";
-import { pinDirectory } from "../pin-directory";
+import { pinFolder } from "../pin-folder";
 import { refreshNumber } from "../hooks/hooks";
 import { ActionRemoveAllDirectories } from "./action-on-files";
 import { DirectoryWithFileInfo } from "../types/types";
@@ -23,7 +23,7 @@ export function QuickAccessEmptyView(props: {
             title={"Pin Directory"}
             icon={Icon.Pin}
             onAction={async () => {
-              await pinDirectory(false);
+              await pinFolder(false);
               setRefresh(refreshNumber());
             }}
           />